Move NInt to its own module, change internal representation to complement
